[+] find person feature [+] PersonProcessor and PersonRepository specific interface for person specific methods [...] Exception handlers moved to interface over controllers Part2
