+++ src/main/java/app/ErrorResponse.java
@@ -3,7 +3,6 @@ package app;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import org.springframework.data.annotation.Id;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,31 +10,19 @@ import java.util.UUID;
 
 @NoArgsConstructor
 @Getter
-public class ErrorResponse{
+public class ErrorResponse {
     private final String id = UUID.randomUUID().toString();
     private final List<ResponseError> errors = new ArrayList<>();
 
-    public ErrorResponse(ResponseError error){
+    public ErrorResponse(ResponseError error) {
         this.getErrors().add(error);
     }
 
-public ErrorResponse(String code, String message){
+    public ErrorResponse(String code, String message) {
         this.errors.add(new ResponseError(code, message));
-}
-
-    @ToString
-    @Getter
-    public static class ResponseError {
-        private final String code;
-        private final String message;
-        @Id
-        private final String id;
-        public ResponseError(String code, String message){
-            this.code = code;
-            this.message = message;
-            this.id = UUID.randomUUID().toString();
-        }
     }
+
+    public record ResponseError(String code, String message) {}
 
     @Override
     public String toString() {
